requirements.txt
pyotp~=2.9
werkzeug~=3.1.3
requests~=2.32.3
pillow~=11.1.0